// store/search/types/enums/mutation-type.ts
export enum MutationType {
    SET_TYPE_MODE_ENABLED = "SET_TYPE_MODE_ENABLED",
    SET_QUERY = "SET_QUERY",
    CLEAR_QUERY = "CLEAR_QUERY",

    SET_LIVE_LOADING = "SET_LIVE_LOADING",
    SET_LIVE_SEARCH_RESULT = "SET_LIVE_SEARCH_RESULT",
    APPEND_LIVE_SEARCH_RESULT = "APPEND_LIVE_SEARCH_RESULT",
    SET_LIVE_SEARCH_DATE = "SET_LIVE_SEARCH_DATE",

    SET_MOVIES_LOADING = "SET_MOVIES_LOADING",
    SET_MOVIES_SEARCH_RESULT = "SET_MOVIES_SEARCH_RESULT",
    APPEND_MOVIES_SEARCH_RESULT = "APPEND_MOVIES_SEARCH_RESULT",

    SET_SHOWS_LOADING = "SET_SHOWS_LOADING",
    SET_SHOWS_SEARCH_RESULT = "SET_SHOWS_SEARCH_RESULT",
    APPEND_SHOWS_SEARCH_RESULT = "APPEND_SHOWS_SEARCH_RESULT",

    SET_CHANNELS_LOADING = "SET_CHANNELS_LOADING",
    SET_CHANNELS_SEARCH_RESULT = "SET_CHANNELS_SEARCH_RESULT",
    APPEND_CHANNELS_SEARCH_RESULT = "APPEND_CHANNELS_SEARCH_RESULT",

    SET_CREDITS_LOADING = "SET_CREDITS_LOADING",
    SET_CREDITS_SEARCH_RESULT = "SET_CREDITS_SEARCH_RESULT",
    APPEND_CREDITS_SEARCH_RESULT = "APPEND_CREDITS_SEARCH_RESULT",

    CLEAR_ALL = "CLEAR_ALL",
    CUT_SEARCH_RESULT = "CUT_SEARCH_RESULT",
    
    SET_LINE_SEARCH_RESULT = "SET_LINE_SEARCH_RESULT",
}